pointerevents/pointerevent_pointercancel_touch.html

<!doctype html>
<html>
    <head>
        <title>PointerCancel - touch</title>
        <meta name="viewport" content="width=device-width">
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
        <!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
    </head>
    <body onload="run()">
        <h1>pointercancel test</h1>
        <h3>Warning: this test works properly only for devices that have touchscreen</h3>
        <h4>Test Description: This test checks if pointercancel event triggers. Start touch over the black rectangle and then move your finger as you are scrolling the page. </h4>
        <p>
        <div id="target0" style="background: black"></div>
        <script>
            var detected_pointertypes = {};
            var test_pointerEvent = async_test("pointercancel event received");
            // showPointerTypes is defined in pointerevent_support.js
            // Requirements: the callbakc function will reference the test_pointerEvent object and
            // will fail unless the async_test is created with the var name "test_pointerEvent".
            add_completion_callback(showPointerTypes);
            
            var pointerdown_event;
            
            function run() {
            	var target0 = document.getElementById("target0");
            
            	on_event(target0, "pointerdown", function (event) {
            		pointerdown_event = event;
            		detected_pointertypes[event.pointerType] = true;
            	});
            
            	on_event(target0, "pointercancel", function (event) {
            		document.getElementById('target0').style.background = 'green';
            
            		test_pointerEvent.step(function () {
            			assert_equals(event.pointerId, pointerdown_event.pointerId, "pointerId should be the same for pointerdown and pointercancel");
            			assert_true(event instanceof PointerEvent,"event is a PointerEvent event");
            			assert_equals(event.pointerType, pointerdown_event.pointerType, "pointerType should be the same for pointerdown and pointercancel");
            			assert_equals(event.isPrimary, pointerdown_event.isPrimary, "isPrimary should be the same for pointerdown and pointercancel");
            			check_PointerEvent(event);
            		});
            
            		detected_pointertypes[event.pointerType] = true;
            
            		test_pointerEvent.done();
            	});
            }
            
        </script>
        <h1>Pointer Events pointercancel Tests</h1>
        <div id="complete-notice">
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
            <p>Refresh the page to run the tests again with a different pointer type.</p>
        </div>
        <div id="log"></div>
    </body>
</html>
